fix merge stuff
